Convert Ints to Longs
Some parts of the Agent send a MAX_INT which in Node is a "Long"

+++ src/main/kotlin/com/sourcegraph/cody/agent/protocol_extensions/Range.kt
@@ -5,7 +5,7 @@ import com.intellij.openapi.editor.Document
 import com.intellij.openapi.editor.RangeMarker
 import com.sourcegraph.cody.agent.protocol_generated.Range
 
-typealias RangePair = Pair<Int, Int>
+typealias RangePair = Pair<Long, Long>
 
 object RangeFactory {
   fun fromRangeMarker(rm: RangeMarker): Range =
@@ -17,7 +17,7 @@ object RangeFactory {
 }
 
 public fun Range.toIntellijRange(): RangePair =
-    RangePair(this.start.line.plus(1), this.end.line.plus(1))
+    RangePair(this.start.line.plus(1L), this.end.line.plus(1L))
 // We need to .plus(1) since the ranges use 0-based indexing
 // but IntelliJ presents it as 1-based indexing.
 public fun Range.intellijRange(): RangePair = RangePair(start.line.plus(1), end.line.plus(1))
